move contracts by type from cli to client
