+++ grid-cli/internal/cmd/get.go
@@ -2,30 +2,20 @@
 package cmd
 
 import (
-	"encoding/json"
 	"fmt"
-	"strconv"
 
 	"github.com/threefoldtech/tfgrid-sdk-go/grid-client/deployer"
 	"github.com/threefoldtech/tfgrid-sdk-go/grid-client/workloads"
 )
 
-var (
-	vmType          = "vm"
-	gatewayNameType = "Gateway Name"
-	gatewayFQDNType = "Gateway Fqdn"
-	k8sType         = "kubernetes"
-	networkType     = "network"
-)
-
 // GetVM gets a vm with its project name
 func GetVM(t deployer.TFPluginClient, name string) (workloads.Deployment, error) {
-	nodeContractIDs, err := getContractsByTypeAndName(t, name, vmType, name)
+	nodeContractIDs, err := t.ContractsGetter.GetNodeContractsByTypeAndName(name, workloads.VMType, name)
 	if err != nil {
 		return workloads.Deployment{}, err
 	}
 
-	networkContractIDs, err := getContractsByTypeAndName(t, name, networkType, fmt.Sprintf("%snetwork", name))
+	networkContractIDs, err := t.ContractsGetter.GetNodeContractsByTypeAndName(name, workloads.NetworkType, fmt.Sprintf("%snetwork", name))
 	if err != nil {
 		return workloads.Deployment{}, err
 	}
@@ -45,12 +35,12 @@ func GetVM(t deployer.TFPluginClient, name string) (workloads.Deployment, error)
 
 // GetK8sCluster gets a kubernetes cluster with its project name
 func GetK8sCluster(t deployer.TFPluginClient, name string) (workloads.K8sCluster, error) {
-	nodeContractIDs, err := getContractsByTypeAndName(t, name, k8sType, name)
+	nodeContractIDs, err := t.ContractsGetter.GetNodeContractsByTypeAndName(name, workloads.K8sType, name)
 	if err != nil {
 		return workloads.K8sCluster{}, err
 	}
 
-	networkContractIDs, err := getContractsByTypeAndName(t, name, networkType, fmt.Sprintf("%snetwork", name))
+	networkContractIDs, err := t.ContractsGetter.GetNodeContractsByTypeAndName(name, workloads.NetworkType, fmt.Sprintf("%snetwork", name))
 	if err != nil {
 		return workloads.K8sCluster{}, err
 	}
@@ -71,7 +61,7 @@ func GetK8sCluster(t deployer.TFPluginClient, name string) (workloads.K8sCluster
 
 // GetGatewayName gets a gateway name with its project name
 func GetGatewayName(t deployer.TFPluginClient, name string) (workloads.GatewayNameProxy, error) {
-	nodeContractIDs, err := getContractsByTypeAndName(t, name, gatewayNameType, name)
+	nodeContractIDs, err := t.ContractsGetter.GetNodeContractsByTypeAndName(name, workloads.GatewayNameType, name)
 	if err != nil {
 		return workloads.GatewayNameProxy{}, err
 	}
@@ -86,7 +76,7 @@ func GetGatewayName(t deployer.TFPluginClient, name string) (workloads.GatewayNa
 
 // GetGatewayFQDN gets a gateway fqdn with its project name
 func GetGatewayFQDN(t deployer.TFPluginClient, name string) (workloads.GatewayFQDNProxy, error) {
-	nodeContractIDs, err := getContractsByTypeAndName(t, name, gatewayFQDNType, name)
+	nodeContractIDs, err := t.ContractsGetter.GetNodeContractsByTypeAndName(name, workloads.GatewayFQDNType, name)
 	if err != nil {
 		return workloads.GatewayFQDNProxy{}, err
 	}
@@ -97,35 +87,4 @@ func GetGatewayFQDN(t deployer.TFPluginClient, name string) (workloads.GatewayFQ
 	}
 
 	return t.State.LoadGatewayFQDNFromGrid(nodeID, name, name)
-}
-
-func getContractsByTypeAndName(t deployer.TFPluginClient, projectName, deploymentType, deploymentName string) (map[uint32]uint64, error) {
-	contracts, err := t.ContractsGetter.ListContractsOfProjectName(projectName)
-	if err != nil {
-		return map[uint32]uint64{}, err
-	}
-	nodeContractIDs := make(map[uint32]uint64)
-	for _, contract := range contracts.NodeContracts {
-		var deploymentData workloads.DeploymentData
-		err := json.Unmarshal([]byte(contract.DeploymentData), &deploymentData)
-		if err != nil {
-			return map[uint32]uint64{}, err
-		}
-		if deploymentData.Type != deploymentType || deploymentData.Name != deploymentName {
-			continue
-		}
-		contractID, err := strconv.ParseUint(contract.ContractID, 0, 64)
-		if err != nil {
-			return map[uint32]uint64{}, err
-		}
-		nodeContractIDs[contract.NodeID] = contractID
-		// only k8s and network have multiple contracts
-		if deploymentType == vmType || deploymentType == gatewayFQDNType || deploymentType == gatewayNameType {
-			break
-		}
-	}
-	if len(nodeContractIDs) == 0 {
-		return map[uint32]uint64{}, fmt.Errorf("no %s with name %s found", deploymentType, deploymentName)
-	}
-	return nodeContractIDs, nil
 }
